+++ src/reader/buffered_reader.rs
@@ -105,7 +105,7 @@ macro_rules! impl_buffered_source {
             buf.push(b'!');
             self $(.$reader)? .consume(1);
 
-            let bang_type = BangType::new(self.peek_one() $(.$await)? ?)?;
+            let bang_type = BangType::new(self.peek_one() $(.$await)? ?, position)?;
 
             loop {
                 match self $(.$reader)? .fill_buf() $(.$await)? {
@@ -139,6 +139,10 @@ macro_rules! impl_buffered_source {
                 }
             }
 
+            // <!....EOF
+            //  ^^^^^ - `buf` does not contains `<`, but we want to report error at `<`,
+            //          so we move offset to it (+1 for `<`)
+            *position -= 1;
             Err(bang_type.to_err())
         }
 
@@ -182,6 +186,10 @@ macro_rules! impl_buffered_source {
                 };
             }
 
+            // <.....EOF
+            //  ^^^^^ - `buf` does not contains `<`, but we want to report error at `<`,
+            //          so we move offset to it (+1 for `<`)
+            *position -= 1;
             Err(Error::Syntax(SyntaxError::UnclosedTag))
         }