+++ src/reader/slice_reader.rs
@@ -286,7 +286,7 @@ impl<'a> XmlSource<'a, ()> for &'a [u8] {
         // start with it.
         debug_assert_eq!(self[0], b'!');
 
-        let bang_type = BangType::new(self[1..].first().copied())?;
+        let bang_type = BangType::new(self[1..].first().copied(), position)?;
 
         if let Some((bytes, i)) = bang_type.parse(&[], self) {
             *position += i;
@@ -294,8 +294,10 @@ impl<'a> XmlSource<'a, ()> for &'a [u8] {
             return Ok((bang_type, bytes));
         }
 
-        // Note: Do not update position, so the error points to
-        // somewhere sane rather than at the EOF
+        // <!....EOF
+        //  ^^^^^ - `self` does not contains `<`, but we want to report error at `<`,
+        //          so we move offset to it (+1 for `<`)
+        *position -= 1;
         Err(bang_type.to_err())
     }
 
@@ -309,8 +311,10 @@ impl<'a> XmlSource<'a, ()> for &'a [u8] {
             return Ok(bytes);
         }
 
-        // Note: Do not update position, so the error points to a sane place
-        // rather than at the EOF.
+        // <.....EOF
+        //  ^^^^^ - `self` does not contains `<`, but we want to report error at `<`,
+        //          so we move offset to it (+1 for `<`)
+        *position -= 1;
         Err(Error::Syntax(SyntaxError::UnclosedTag))
     }